add models for equipment maintenance micro service

+++ src/server.js
@@ -13,8 +13,10 @@ const startServer = async () => {
     console.log('Models are loaded successfully');
     await sequelize.authenticate();
     console.log('Database connected successfully');
-    await sequelize.sync({ force: true });
-    console.log('Database synced successfully');
+    if (config.ENABLE_MODEL_SYNC) {
+      await sequelize.sync({ force: true });
+      console.log('Database synced successfully');
+    }
 
     getRouter().then((router) => {
       app.use('/', router);
